day 20

+++ src/main/kotlin/com/ngr/aoc/y2022/day20/Item.kt
@@ -1,3 +1,3 @@
 package com.ngr.aoc.y2022.day20
 
-data class Item(val value: Int, val order: Int)
+data class Item(val value: Long, val order: Int)
